+++ modules/services/analytics/src/main/java/au/com/shiftyjelly/pocketcasts/analytics/AnalyticsEvent.kt
@@ -515,9 +515,16 @@ enum class AnalyticsEvent(val key: String) {
     SETTINGS_STORAGE_WARN_BEFORE_USING_DATA_TOGGLED("settings_storage_warn_before_using_data_toggled"),
 
     /* Search */
+    SEARCH_SHOWN("search_shown"),
+    SEARCH_DISMISSED("search_dismissed"),
     SEARCH_PERFORMED("search_performed"),
     SEARCH_FAILED("search_failed"),
     SEARCH_RESULT_TAPPED("search_result_tapped"),
+
+    /* Search History */
+    SEARCH_HISTORY_CLEARED("search_history_cleared"),
+    SEARCH_HISTORY_ITEM_TAPPED("search_history_item_tapped"),
+    SEARCH_HISTORY_ITEM_DELETE_BUTTON_TAPPED("search_history_item_delete_button_tapped"),
 
     /* Chromecast */
     CHROMECAST_VIEW_SHOWN("chromecast_view_shown"),