+++ modules/services/model/src/main/java/au/com/shiftyjelly/pocketcasts/models/db/AppDatabase.kt
@@ -25,6 +25,7 @@ import au.com.shiftyjelly.pocketcasts.models.db.dao.EpisodeDao
 import au.com.shiftyjelly.pocketcasts.models.db.dao.FolderDao
 import au.com.shiftyjelly.pocketcasts.models.db.dao.PlaylistDao
 import au.com.shiftyjelly.pocketcasts.models.db.dao.PodcastDao
+import au.com.shiftyjelly.pocketcasts.models.db.dao.SearchHistoryDao
 import au.com.shiftyjelly.pocketcasts.models.db.dao.UpNextChangeDao
 import au.com.shiftyjelly.pocketcasts.models.db.dao.UpNextDao
 import au.com.shiftyjelly.pocketcasts.models.db.dao.UserEpisodeDao
@@ -34,6 +35,7 @@ import au.com.shiftyjelly.pocketcasts.models.entity.Folder
 import au.com.shiftyjelly.pocketcasts.models.entity.Playlist
 import au.com.shiftyjelly.pocketcasts.models.entity.PlaylistEpisode
 import au.com.shiftyjelly.pocketcasts.models.entity.Podcast
+import au.com.shiftyjelly.pocketcasts.models.entity.SearchHistoryItem
 import au.com.shiftyjelly.pocketcasts.models.entity.UpNextChange
 import au.com.shiftyjelly.pocketcasts.models.entity.UpNextEpisode
 import au.com.shiftyjelly.pocketcasts.models.entity.UserEpisode
@@ -48,11 +50,12 @@ import au.com.shiftyjelly.pocketcasts.localization.R as LR
         Playlist::class,
         PlaylistEpisode::class,
         Podcast::class,
+        SearchHistoryItem::class,
         UpNextChange::class,
         UpNextEpisode::class,
         UserEpisode::class,
     ],
-    version = 73,
+    version = 74,
     exportSchema = true
 )
 @TypeConverters(
@@ -77,6 +80,7 @@ abstract class AppDatabase : RoomDatabase() {
     abstract fun userEpisodeDao(): UserEpisodeDao
     abstract fun folderDao(): FolderDao
     abstract fun bumpStatsDao(): BumpStatsDao
+    abstract fun searchHistoryDao(): SearchHistoryDao
 
     companion object {
         // This seems dodgy but I got it from Google, https://github.com/googlesamples/android-sunflower/blob/master/app/src/main/java/com/google/samples/apps/sunflower/data/AppDatabase.kt
@@ -374,6 +378,35 @@ abstract class AppDatabase : RoomDatabase() {
                     );
                 """.trimIndent()
             )
+        }
+
+        val MIGRATION_73_74 = addMigration(73, 74) { database ->
+            database.execSQL(
+                """
+                    CREATE TABLE IF NOT EXISTS search_history (
+                        _id INTEGER PRIMARY KEY AUTOINCREMENT,
+                        modified INTEGER NOT NULL,
+                        term TEXT,
+                        podcast_uuid TEXT,
+                        podcast_title TEXT,
+                        podcast_author TEXT,
+                        folder_uuid TEXT,
+                        folder_title TEXT,
+                        folder_color INTEGER,
+                        folder_podcastIds TEXT,
+                        episode_uuid TEXT,
+                        episode_title TEXT,
+                        episode_duration REAL,
+                        episode_podcastUuid TEXT, 
+                        episode_podcastTitle TEXT, 
+                        episode_artworkUrl TEXT
+                    );
+                """.trimIndent()
+            )
+            database.execSQL("CREATE UNIQUE INDEX IF NOT EXISTS `index_search_history_term` ON search_history (`term`)")
+            database.execSQL("CREATE UNIQUE INDEX IF NOT EXISTS `index_search_history_podcast_uuid` ON search_history (`podcast_uuid`);")
+            database.execSQL("CREATE UNIQUE INDEX IF NOT EXISTS `index_search_history_folder_uuid` ON search_history (`folder_uuid`)")
+            database.execSQL("CREATE UNIQUE INDEX IF NOT EXISTS `index_search_history_episode_uuid` ON search_history (`episode_uuid`)")
         }
 
         fun addMigrations(databaseBuilder: Builder<AppDatabase>, context: Context) {
@@ -738,7 +771,8 @@ abstract class AppDatabase : RoomDatabase() {
                 MIGRATION_69_70,
                 MIGRATION_70_71,
                 MIGRATION_71_72,
-                MIGRATION_72_73
+                MIGRATION_72_73,
+                MIGRATION_73_74
             )
         }